Added a filter feature for data stores
This push is a big one that creates a query builder based on the different data stores. some data stores are not included with this update but others will be added. I may decide to allow the user the ability to add a custom filter handler but not any time soon.